+++ Geoscape/RecruitInfo/changelog.md
@@ -2,7 +2,8 @@ Changelog of Recruit Info, a Phoenix Point Mod by Sheepy
 
 # Version 2
 
-* New: Config to change font size
+* New: Config to change font size and what info to display.
+* Fix: PPML 0.1 compatibility
 
 # Version 1, 2019-12-27